Replace chroot based debootstrap with QEMU
To better support foreign debootstraps, and to debootstrap hermetically,
move away from using chroots. The chroot feature for foreign archs needs
qemu-user, and some versions of these tools crash randomly, which has
caused problems generating new filesystems.

This new method is quite a bit more complicated, but it also provides a
lot more validation. The user must specify a prebuilt kernel and
initramfs, and the initramfs is combined with one produced by the stage2
script, which allows the suite script to run under the kernel it will be
used with, which guarantees the combination will work.

Since we require the user to specify a kernel, we can also support
embedding the kernel, ramdisk (and for arm64 dtb) into the rootfs at
build time, because `uname -r` expands to the true kernel that will be
used.

The script supports building filesystems for non-virtual devices, but
the kernel provided must have at least virtio pci, blk, net and rng to
complete successfully.

These changes also finally enable a long requested capability to boot
the net tests root filesystem with a modular kernel, so that the GKI
kernel prebuilts can be used directly against it, without having to
build a custom "built-in" kernel equivalent. This in turn will enable
changes to run_net_test.sh to skip the kernel build and use cuttlefish
to host the test, which lets us integrate the testing with ATP.

Finally, these changes have been used to build Other OS root filesystems
for cuttlefish, and the root filesystem for the RockPi devices we are
using for ARM64 lab testing.

The bootstrapping process is conducted in three stages:

1) The debootstrap tool is run in --foreign mode offline to download
   the needed packages for installation. However, no installation is
   done and the binaries are simply unpacked, which works without
   needing qemu-user to be installed.

2) QEMU is used to boot the user-provided kernel and a basic root
   filesystem from debootstrap with a stage1.sh control script. The root
   filesystem (512M max) is passed as a ramdisk, so no special drivers
   are required. The stage1.sh script loads some virtual machine modules
   and sets up the psuedo-filesystems, like the old debootstrap second
   stage chroot did. Then, chains to stage2.sh to complete the
   debootstrap second stage on *another* copy of the root filesystem
   from virtio_blk. This process generates an initrd fragment which can
   be combined with the user's initramfs to boot the final system.

3) Finally, the "installed" root filesystem is booted with the Debian
   initial ramdisk and user-provided kernel modules. Once the root
   filesystem is mounted, the suite script runs ("stage3"). This script
   makes the usual customizations from our original chroot method.

If the -e option is specified, the kernel and "baked" initramfs are
stored in /boot. This process works for cuttlefish and rockpi but for
the normal net tests setup, embedding is not preferred (as the kernel
will always be changing). The grub2-common package is also installed to
create a grub.cfg for bootloader config chaining on cuttlefish.

+++ net/test/rootfs/common.sh
@@ -15,6 +15,8 @@
 # limitations under the License.
 #
 
+trap "echo 3 >${exitcode}" ERR
+
 # $1 - Suite name for apt sources
 update_apt_sources() {
   # Add the needed debian sources
@@ -49,6 +51,11 @@ remove_installed_packages() {
 }
 
 setup_static_networking() {
+  # Temporarily bring up static QEMU SLIRP networking (no DHCP)
+  ip link set dev eth0 up
+  ip addr add 10.0.2.15/24 broadcast 10.0.2.255 dev eth0
+  ip route add default via 10.0.2.2 dev eth0
+
   # Permanently update the resolv.conf with the Google DNS servers
   echo "nameserver 8.8.8.8"  >/etc/resolv.conf
   echo "nameserver 8.8.4.4" >>/etc/resolv.conf
@@ -106,6 +113,23 @@ create_systemd_getty_symlinks() {
   done
 }
 
+# $1 - Additional default command line
+setup_grub() {
+  if [ -n "${embed_kernel_initrd_dtb}" ]; then
+    # For testing the image with a virtual device
+    apt-get install -y grub2-common
+    cat >/etc/default/grub <<EOF
+GRUB_DEFAULT=0
+GRUB_TIMEOUT=5
+GRUB_DISTRIBUTOR=Debian
+GRUB_CMDLINE_LINUX_DEFAULT="quiet"
+GRUB_CMDLINE_LINUX="\\\$cmdline $1"
+EOF
+    mkdir /boot/grub
+    update-grub
+  fi
+}
+
 cleanup() {
   # Prevents systemd boot issues with read-only rootfs
   mkdir -p /var/lib/systemd/{coredump,linger,rfkill,timesync}
@@ -114,10 +138,17 @@ cleanup() {
   # If embedding isn't enabled, remove the embedded modules and initrd and
   # uninstall the tools to regenerate the initrd, as they're unlikely to
   # ever be used
-  apt-get purge -y initramfs-tools initramfs-tools-core klibc-utils kmod
+  if [ -z "${embed_kernel_initrd_dtb}" ]; then
+    apt-get purge -y initramfs-tools initramfs-tools-core klibc-utils kmod
+    rm -f "/boot/initrd.img-$(uname -r)"
+    rm -rf "/lib/modules/$(uname -r)"
+  fi
 
   # Miscellaneous cleanup
   rm -rf /var/lib/apt/lists/* || true
   rm -f /root/* || true
   apt-get clean
+
+  echo 0 >"${exitcode}"
+  sync && poweroff -f
 }
